+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog of lizard-fewsjdbc
 1.16 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Added IconStyle model and migration. Icons are now configurable. It
+  will revert to a default when nothing is configurated.
 
 
 1.15 (2011-07-26)